+++ smart_fruit/feature_types/__init__.py
@@ -1,4 +1,4 @@
 from smart_fruit.feature_types.feature_type_base import FeatureType
-from smart_fruit.feature_types.simple_types import Number, Integer, Label
+from smart_fruit.feature_types.simple_types import Number, Integer, Complex, Label
 
-__all__ = ["FeatureType", "Number", "Integer", "Label"]
+__all__ = ["FeatureType", "Number", "Integer", "Complex", "Label"]